Fix various compiler warnings

+++ src/hardware/devices/vga_attrctrl.cpp
@@ -97,7 +97,7 @@ const char * VGA_AttrCtrl::register_to_string(uint8_t _index) const
 
 void VGA_AttrCtrl::registers_to_textfile(FILE *_file)
 {
-	fprintf(_file, "     0x%02X %*u  Address [%s]\n", address, 3, address, (const char*)address);
+	fprintf(_file, "     0x%02X %*u  Address [%s]\n", uint8_t(address), 3, uint8_t(address), (const char*)address);
 	for(int i=0; i<ATTC_REGCOUNT; i++) {
 		fprintf(_file, "0x%02X 0x%02X %*u  %s\n", i, get_register(i), 3, get_register(i), register_to_string(i));
 	}
